解决多个gateway的clientid在game中冲突，调整了对象的消息广播;支持10组机器人;服务器后台信息完善

+++ server/GameServer/GameGatewayMgr.h
@@ -33,7 +33,7 @@ public:
 	// 处理服务器发来的消息
 	void ProcessMsg(serverinfo *info);
 	// 处理Client发来的消息，此时Client已经验证成功了
-	void ProcessClientMsg(int32 clientid, Msg *pMsg);
+	void ProcessClientMsg(int64 clientid, Msg *pMsg);
 
 private:
 	// 添加新的Server
